requirements.txt
requests
pydantic
typing-extensions
llama-index-core
llama-index-llms-openai
llama-index-llms-huggingface
python-dotenv